+++ front/app/containers/IdeasEditPageOld/messages.ts
@@ -1,0 +1,54 @@
+import { defineMessages } from 'react-intl';
+
+export default defineMessages({
+  formTitle: {
+    id: 'app.containers.IdeasEditPage.formTitle',
+    defaultMessage: 'Edit idea',
+  },
+  optionFormTitle: {
+    id: 'app.containers.IdeasEditPage.optionFormTitle',
+    defaultMessage: 'Edit option',
+  },
+  projectFormTitle: {
+    id: 'app.containers.IdeasEditPage.projectFormTitle',
+    defaultMessage: 'Edit project',
+  },
+  questionFormTitle: {
+    id: 'app.containers.IdeasEditPage.questionFormTitle',
+    defaultMessage: 'Edit question',
+  },
+  issueFormTitle: {
+    id: 'app.containers.IdeasEditPage.issueFormTitle',
+    defaultMessage: 'Edit issue',
+  },
+  contributionFormTitle: {
+    id: 'app.containers.IdeasEditPage.contributionFormTitle',
+    defaultMessage: 'Edit contribution',
+  },
+  save: {
+    id: 'app.containers.IdeasEditPage.save',
+    defaultMessage: 'Save',
+  },
+  submitApiError: {
+    id: 'app.containers.IdeasEditPage.submitApiError',
+    defaultMessage:
+      'There was an issue submitting the form. Please check for any errors and try again.',
+  },
+  editedPostSave: {
+    id: 'app.containers.IdeasEditPage.editedPostSave',
+    defaultMessage: 'Save',
+  },
+  fileUploadError: {
+    id: 'app.containers.IdeasEditPage.fileUploadError',
+    defaultMessage:
+      'One or more files failed to upload. Please check the file size and format and try again.',
+  },
+  ideasEditMetaTitle: {
+    id: 'app.containers.IdeasEditPage.ideasEditMetaTitle',
+    defaultMessage: 'Edit {postTitle} | {projectName}',
+  },
+  ideasEditMetaDescription: {
+    id: 'app.containers.IdeasEditPage.ideasEditMetaDescription',
+    defaultMessage: 'Edit your post. Add new and change old information.',
+  },
+});
