+++ front/app/containers/IdeasNewPageOld/tracks.ts
@@ -1,0 +1,8 @@
+export default {
+  titleProfanityError: {
+    name: 'Profane idea title blocked',
+  },
+  descriptionProfanityError: {
+    name: 'Profane idea description blocked',
+  },
+};
