+++ front/app/hooks/useInputSchema.ts
@@ -1,0 +1,4 @@
+export default (_projectId) => ({
+  schema: {},
+  uiSchema: {},
+});
